CARL ResourceManager: Move `storage` underneath `persistence`.
